# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "L.U.C.I.A. API",
    "description": "Logic & Ultimate Chess Insight Algorithm",
    "version": "0.2.5"
  },
  "paths": {
...
    "/analysis": {
      "post": {
        "tags": [
          "analysis"
        ],
        "summary": "Create Analysis",
        "operationId": "create_analysis_analysis_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AnalysisSummary"
                  },
                  "title": "Response Create Analysis Analysis Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "analysis"
        ],
        "summary": "List Analyses",
        "description": "Análisis existentes, opcionalmente los de una partida o un tablero\nconcretos. Sirve para que el visor —y el tablero de análisis (RF-6.9)—\nsepan si ya hay uno hecho en vez de volver a analizar.",
        "operationId": "list_analyses_analysis_get",
        "parameters": [
          {
            "name": "game_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Game Id"
            }
          },
          {
            "name": "board_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Board Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AnalysisSummary"
                  },
                  "title": "Response List Analyses Analysis Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analysis/position": {
      "post": {
        "tags": [
          "analysis"
        ],
        "summary": "Position Analysis",
        "description": "Analiza una posición suelta y devuelve las mejores líneas (RF-5.2 /\nRF-6.2). Para el tablero de análisis y la exploración en vivo.",
        "operationId": "position_analysis_analysis_position_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PositionAnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EngineLineOut"
                  },
                  "type": "array",
                  "title": "Response Position Analysis Analysis Position Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analysis/compare": {
      "get": {
        "tags": [
          "analysis"
        ],
        "summary": "Compare",
        "description": "Dónde discrepan dos análisis de la misma partida (RF-2.6), típicamente\nuno de Stockfish y otro de Lc0.",
        "operationId": "compare_analysis_compare_get",
        "parameters": [
          {
            "name": "analysis_a",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Analysis A"
            }
          },
          {
            "name": "analysis_b",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Analysis B"
            }
          },
          {
            "name": "threshold",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "maximum": 100,
              "minimum": 0,
              "default": 10.0,
              "title": "Threshold"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisComparisonOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analysis/{analysis_id}": {
      "get": {
        "tags": [
          "analysis"
        ],
        "summary": "Get Analysis",
        "operationId": "get_analysis_analysis__analysis_id__get",
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Analysis Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analysis/{analysis_id}/pgn": {
      "get": {
        "tags": [
          "analysis"
        ],
        "summary": "Get Analysis Pgn",
        "description": "El PGN de la partida con las anotaciones del análisis (RF-5.5).\n\nSe descarga como archivo —de ahí el `Content-Disposition`— porque lo que\nse quiere hacer con él es abrirlo en lichess, ChessBase o SCID, no leerlo\nen el navegador.\n\nSolo se exporta un análisis **terminado**: uno a medias daría una partida\ncomentada hasta la jugada 20 y muda a partir de ahí, que se lee como un\narchivo roto y no como un análisis en curso.",
        "operationId": "get_analysis_pgn_analysis__analysis_id__pgn_get",
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Analysis Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AnalysisComparisonOut": {
        "properties": {
          "game_id": {
            "type": "integer",
            "title": "Game Id"
          },
          "analysis_a": {
            "type": "integer",
            "title": "Analysis A"
          },
          "analysis_b": {
            "type": "integer",
            "title": "Analysis B"
          },
          "engine_a": {
            "type": "string",
            "title": "Engine A"
          },
          "engine_b": {
            "type": "string",
            "title": "Engine B"
          },
          "total_moves": {
            "type": "integer",
            "title": "Total Moves"
          },
          "agreed_best_moves": {
            "type": "integer",
            "title": "Agreed Best Moves"
          },
          "best_move_agreement_percent": {
            "type": "number",
            "title": "Best Move Agreement Percent"
          },
          "disagreements": {
            "items": {
              "$ref": "#/components/schemas/MoveComparisonOut"
            },
            "type": "array",
            "title": "Disagreements"
          }
        },
        "type": "object",
        "required": [
          "game_id",
          "analysis_a",
          "analysis_b",
          "engine_a",
          "engine_b",
          "total_moves",
          "agreed_best_moves",
          "best_move_agreement_percent",
          "disagreements"
        ],
        "title": "AnalysisComparisonOut"
      },
      "AnalysisDetail": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "game_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Game Id"
          },
          "board_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Board Id"
          },
          "engine": {
            "type": "string",
            "title": "Engine"
          },
          "depth": {
            "type": "integer",
            "title": "Depth"
          },
          "multipv": {
            "type": "integer",
            "title": "Multipv"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "white_accuracy": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "White Accuracy"
          },
          "black_accuracy": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Black Accuracy"
          },
          "moves": {
            "items": {
              "$ref": "#/components/schemas/AnalyzedMoveOut"
            },
            "type": "array",
            "title": "Moves"
          },
          "critical_moments": {
            "items": {
              "$ref": "#/components/schemas/CriticalMomentOut"
            },
            "type": "array",
            "title": "Critical Moments",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "game_id",
          "board_id",
          "engine",
          "depth",
          "multipv",
          "status",
          "error",
          "white_accuracy",
          "black_accuracy",
          "moves"
        ],
        "title": "AnalysisDetail"
      },
      "AnalysisRequest": {
        "properties": {
          "game_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Game Ids"
          },
          "engine": {
            "type": "string",
            "enum": [
              "stockfish",
              "lc0"
            ],
            "title": "Engine",
            "default": "stockfish"
          },
          "depth": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Depth"
          },
          "multipv": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Multipv"
          }
        },
        "type": "object",
        "required": [
          "game_ids"
        ],
        "title": "AnalysisRequest"
      },
      "AnalysisSummary": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "game_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Game Id"
          },
          "board_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Board Id"
          },
          "engine": {
            "type": "string",
            "title": "Engine"
          },
          "depth": {
            "type": "integer",
            "title": "Depth"
          },
          "multipv": {
            "type": "integer",
            "title": "Multipv"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "white_accuracy": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "White Accuracy"
          },
          "black_accuracy": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Black Accuracy"
          }
        },
        "type": "object",
        "required": [
          "id",
          "game_id",
          "board_id",
          "engine",
          "depth",
          "multipv",
          "status",
          "error",
          "white_accuracy",
          "black_accuracy"
        ],
        "title": "AnalysisSummary"
      },
...
      "EngineLineOut": {
        "properties": {
          "rank": {
            "type": "integer",
            "title": "Rank"
          },
          "score_cp": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score Cp"
          },
          "score_mate": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score Mate"
          },
          "pv_uci": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Pv Uci"
          },
          "pv_san": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Pv San"
          }
        },
        "type": "object",
        "required": [
          "rank",
          "score_cp",
          "score_mate",
          "pv_uci",
          "pv_san"
        ],
        "title": "EngineLineOut",
        "description": "Una línea del motor, igual la calcule en vivo `POST /analysis/position`\no venga guardada con una jugada analizada (RF-10.1): mismo formato en los\ndos sitios, para que el front la dibuje igual."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PositionAnalysisRequest": {
        "properties": {
          "fen": {
            "type": "string",
            "title": "Fen"
          },
          "engine": {
            "type": "string",
            "enum": [
              "stockfish",
              "lc0"
            ],
            "title": "Engine",
            "default": "stockfish"
          },
          "depth": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 22.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Depth"
          },
          "multipv": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 10.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Multipv"
          }
        },
        "type": "object",
        "required": [
          "fen"
        ],
        "title": "PositionAnalysisRequest"
      },
...
    }
  }
}